assignments/W3/exercises/requirements.txt
# Requirements for Prompt Engineering Basics script
# Install with: pip install -r requirements.txt

transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
accelerate>=0.20.0

